config-scope-resolution: integration-bucket residual closure + fixture migration
Carries the config-scope-resolution working set to a green tree. The
integration bucket is now reachable end to end (it was previously masked by
earlier failing binaries), and its four residual reds are closed:

- perimeter_parity::annulus_true_hole_produces_inner_perimeters: the test's
  hard-coded 6-wall expectation predates 3904c361, which aligned the
  wall_loops default to canonical `wall_loops = 2` (verified against
  PrintConfigDef::init_fff_params). The module fallback of 3 was the
  pre-rename internal inconsistency. Authoring wall_loops = 3 restores the
  test's intended 3 outer + 3 hole loops. The 15 boostvoronoi
  `is_finite` panics that accompany the failing run are DEV-093's
  caught-and-converted panics; they are absent once the wall count is 3.
- no_linker_module_degraded_raw_output: the block-based discriminator went
  stale when packets 233/234/235 reshaped the wedge's sparse-infill islands
  (measured 30.93 against a 28.0 threshold, without the linker). Replaced
  with the AC's literal unit, mean points per sparse-infill path, measured
  on this tree: 2.00 without the linker vs 4.90 with it (threshold 2.5).
- final_gcode_roles (interface_layer_count_follows_config): packet-04's
  automatic-value expansion mirrors a raw support_interface_bottom_layers
  = -1 to the top count (canonical
  number_of_support_interface_bottom_layers), so -1 must equal an explicit
  bottom = top run rather than bottom = 0. The packet-238c G-18 widening
  legitimately adds one top-band layer for a positive bottom count; the
  green interface_band_counts_match_canonical_structure pin is preserved.
- pnp-cli::missing_layer_collection_geometry_fails_with_render_failed_via_cli:
  packet-04 expansion requires a nozzle_diameter-declaring module to expand
  the host default line_width = 0; the minimal visual-debug module set now
  includes machine-gcode-emit so the test reaches its RenderFailed
  assertion instead of failing earlier with CaptureFailed.

Also in this commit:

- slicer-config: registry assembly + resolution/expansion refinements
  (config_block_map_for_modules, BOTTOM_AUTO_BASES -1 mirrors, parse_finite
  percent spelling).
- slicer-ir: line_width f32 -> f64 typed field; resolved-config defaults.
- slicer-runtime: typed module config source / overlay-expanded global,
  layer-executor and prepass support plumbing.
- slicer-gcode: CONFIG_BLOCK float rendering precision.
- slicer-scheduler: execution-plan and region-split updates.
- slicer-wasm-host: dispatch and marshal adjustments.
- core-modules: top-surface-ironing ironing_flow percent migration
  (0..100), tree-support-planner raft_first_layer_density percent migration
  (10..100), matching manifest and doc updates.
- 21 integration/contract/executor/e2e test targets migrated to the
  production-resolved bound-view baseline pattern
  (classic_perimeters_baseline) and registry-typed defaults.
- Regenerated docs/07 Open Deviation Map and docs/15 config tables
  (check-deviations --check is green).

Gates: cargo clippy --workspace --all-targets -- -D warnings; cargo xtask
check-literals; cargo xtask build-guests --check; cargo run -p xtask --
check-deviations --check; rustfmt --check over all changed files.

+++ crates/slicer-core/src/algos/lightning/mod.rs
@@ -88,7 +88,7 @@ pub fn generate_lightning_trees(
                 false,
                 false,
                 &crate::flow::RoleWidthContext {
-                    line_width: config.line_width,
+                    line_width: config.line_width as f32,
                     nozzle_diameter: DEFAULT_NOZZLE_DIAMETER_MM,
                     ..Default::default()
                 },
